DS-2077 Sort collections by full path
This commit fixes the order of collection drop-down entries when starting a
submission and when editing e-mail subscriptions such that the community
hierarchy is taken into account as well as the collection name. The change in
CollectionDropDown makes the same mechanism re-usable in other places where
this might be required.